feat(android): add native grabber view with JS hitbox for drag handling
