+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from './TrueSheet';
 export * from './TrueSheet.types';
-export * from './TrueSheetGrabber';
 export * from './reanimated';
